+++ releases/latest/CHANGELOG.md
@@ -1,6 +1,6 @@
 ### [0.5.6]
 
-
+- Added the option to disable the quota from the LethalConfig menu
 
 
 ### [0.5.5]